+++ src/aplication/Routes.js
@@ -1,7 +1,8 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom"
-import Footer from "../components/Footer";
-import Nav from "../components/Nav";
-import Home from "../pages/Home"
+import Footer from "../components/footer/Footer";
+import Nav from "../components/nav/Nav";
+import Home from "../pages/home/Home"
+import Food from "../pages/food/Food"
 
 //import PrivateRoute from "./privateRoute";
 
@@ -13,6 +14,7 @@ const Router = () => {
             <Routes>
                 <Route path={process.env.PUBLIC_URL}>
                     <Route index element={<Home />} />
+                    <Route path="select/" element={<Food />} />
                 </Route>
                 <Route path="*" element={<div><h2>404 not found</h2></div>} />
             </Routes>
