+++ src/components/footer/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import FooterStyle from './FooterStyle'
+import FooterStyle from '../footer/FooterStyle'
 
 const Footer = () => {
     return (
