+++ src/components/nav/NavStyle.js
@@ -3,9 +3,11 @@ import styled from "styled-components";
 const NavStyle = styled.div`
 
     display: flex;
-    justify-content: space-between ;
-    padding: 8vh; 
+    flex-direction: column;
+    padding: 5vh; 
     background-color: #6ef1d0d5;
+
+    
 `
 
 export default NavStyle;
