+++ src/components/welcome/Welcome.js
@@ -1,6 +1,10 @@
 import React from 'react'
 
 const Welcome = () => {
+
+
+
+
     return (
         <div>Welcome</div>
     )
